Fix '--help' flag not ending program execution
Specifically, prevent a use case where flags such as '--help' and
'--version' that should quit the program immediately after printing a
message were not doing so. This meant that an error message was also
printed if the user had provided invalid input to specify the dice.

An example scenario where this would happen: 'dice --help coffee'